:duck: add animals component
