+++ src/components/MenuBar.vue
@@ -2,7 +2,8 @@
   <div class="level is-mobile">
     <div class="level-item has-text-centered">
       <NuxtLink to="/" class="item button is-primary is-outlined">
-        <canvas ref="canvasNeko" class="neko" width="40" height="40" /><span
+        <MenuNeko />
+        <span
           class="lnk"
           >Home</span
         >
@@ -10,7 +11,8 @@
     </div>
     <div class="level-item">
       <NuxtLink to="/about" class="item button is-primary is-outlined">
-        <canvas ref="canvasKuma" class="kuma" width="40" height="40" /><span
+        <MenuKuma />
+        <span
           class="lnk"
           >About</span
         >
@@ -19,7 +21,8 @@
 
     <div class="level-item">
       <NuxtLink to="/works" class="item button is-primary is-outlined">
-        <canvas ref="canvasUsagi" class="usagi" width="40" height="40" /><span
+        <MenuUsagi />
+        <span
           class="lnk"
           >Works</span
         >
@@ -27,7 +30,8 @@
     </div>
     <div class="level-item">
       <NuxtLink to="/contact" class="item button is-primary is-outlined">
-        <canvas ref="canvasSakana" class="sakana" width="40" height="40" /><span
+        <MenuSakana />
+        <span
           class="lnk"
           >Contact</span
         >
@@ -35,50 +39,6 @@
     </div>
   </div>
 </template>
-
-<script setup>
-import { makeNeko } from "~/zdog/neko.js"
-import { makeKuma } from "~/zdog/kuma.js"
-import { makeUsagi } from "~/zdog/usagi.js"
-import { makeSakana } from "~/zdog/sakana.js"
-
-const app = useNuxtApp()
-
-const canvasNeko = ref()
-const neko = ref()
-const canvasKuma = ref()
-const kuma = ref()
-const canvasUsagi = ref()
-const usagi = ref()
-const canvasSakana = ref()
-const sakana = ref()
-
-watchEffect(() => {
-  if (neko.value) {
-    neko.value.rotate.set(app.$rotate)
-    neko.value.updateRenderGraph()
-  }
-  if (kuma.value) {
-    kuma.value.rotate.set(app.$rotate)
-    kuma.value.updateRenderGraph()
-  }
-  if (usagi.value) {
-    usagi.value.rotate.set(app.$rotate)
-    usagi.value.updateRenderGraph()
-  }
-  if (sakana.value) {
-    sakana.value.rotate.set(app.$rotate)
-    sakana.value.updateRenderGraph()
-  }
-})
-
-onMounted(() => {
-  neko.value = makeNeko(canvasNeko.value, 0.14, true)
-  kuma.value = makeKuma(canvasKuma.value, 0.14, true)
-  usagi.value = makeUsagi(canvasUsagi.value, 0.14, true)
-  sakana.value = makeSakana(canvasSakana.value, 0.14, true)
-})
-</script>
 
 <style css scoped>
 .item {
